requirements.txt
flask
gunicorn
gevent
flask-socketio
gevent-websocket
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
celery
redis
requests
playwright
sqlalchemy-stubs
types-redis
psycogreen
google-generativeai
praw==7.7.1
httpx==0.27.0
google-cloud-secret-manager==2.20.0